# Exclude load paths