--- a/sorbet/rbi/gems/parallel@1.25.1.rbi
+++ b/sorbet/rbi/gems/parallel@1.25.1.rbi
@@ -24,10 +24,10 @@ module Parallel
     def each_with_index(array, options = T.unsafe(nil), &block); end
 
     # source://parallel//lib/parallel.rb#307
-    def filter_map(*args, &block); end
+    def filter_map(*_arg0, **_arg1, &_arg2); end
 
     # source://parallel//lib/parallel.rb#303
-    def flat_map(*args, &block); end
+    def flat_map(*_arg0, **_arg1, &_arg2); end
 
     # source://parallel//lib/parallel.rb#228
     def in_processes(options = T.unsafe(nil), &block); end
@@ -48,68 +48,74 @@ module Parallel
 
     # Number of processors seen by the OS, used for process scheduling
     #
-    # source://parallel//lib/parallel.rb#345
+    # source://parallel//lib/parallel.rb#341
     def processor_count; end
 
-    # source://parallel//lib/parallel.rb#350
+    # source://parallel//lib/parallel.rb#346
     def worker_number; end
 
     # TODO: this does not work when doing threads in forks, so should remove and yield the number instead if needed
     #
-    # source://parallel//lib/parallel.rb#355
+    # source://parallel//lib/parallel.rb#351
     def worker_number=(worker_num); end
 
     private
 
-    # source://parallel//lib/parallel.rb#361
+    # source://parallel//lib/parallel.rb#384
     def add_progress_bar!(job_factory, options); end
 
-    # source://parallel//lib/parallel.rb#624
+    # source://parallel//lib/parallel.rb#647
     def call_with_index(item, index, options, &block); end
 
-    # source://parallel//lib/parallel.rb#556
+    # source://parallel//lib/parallel.rb#579
     def create_workers(job_factory, options, &block); end
 
     # options is either a Integer or a Hash with :count
     #
-    # source://parallel//lib/parallel.rb#614
+    # source://parallel//lib/parallel.rb#637
     def extract_count_from_options(options); end
 
-    # source://parallel//lib/parallel.rb#642
+    # source://parallel//lib/parallel.rb#665
     def instrument_finish(item, index, result, options); end
 
     # yield results in the order of the input items
     # needs to use `options` to store state between executions
     # needs to use `done` index since a nil result would also be valid
     #
-    # source://parallel//lib/parallel.rb#651
+    # source://parallel//lib/parallel.rb#674
     def instrument_finish_in_order(item, index, result, options); end
 
-    # source://parallel//lib/parallel.rb#671
+    # source://parallel//lib/parallel.rb#694
     def instrument_start(item, index, options); end
 
-    # source://parallel//lib/parallel.rb#590
+    # source://parallel//lib/parallel.rb#357
+    def physical_processor_count_windows; end
+
+    # source://parallel//lib/parallel.rb#613
     def process_incoming_jobs(read, write, job_factory, options, &block); end
 
-    # source://parallel//lib/parallel.rb#544
+    # source://parallel//lib/parallel.rb#567
     def replace_worker(job_factory, workers, index, options, blk); end
 
-    # source://parallel//lib/parallel.rb#635
+    # source://parallel//lib/parallel.rb#378
+    def run(command); end
+
+    # source://parallel//lib/parallel.rb#658
     def with_instrumentation(item, index, options); end
 
-    # source://parallel//lib/parallel.rb#386
+    # source://parallel//lib/parallel.rb#409
     def work_direct(job_factory, options, &block); end
 
-    # source://parallel//lib/parallel.rb#496
+    # source://parallel//lib/parallel.rb#519
     def work_in_processes(job_factory, options, &blk); end
 
-    # source://parallel//lib/parallel.rb#430
+    # source://parallel//lib/parallel.rb#453
     def work_in_ractors(job_factory, options); end
 
-    # source://parallel//lib/parallel.rb#405
+    # source://parallel//lib/parallel.rb#428
     def work_in_threads(job_factory, options, &block); end
 
-    # source://parallel//lib/parallel.rb#564
+    # source://parallel//lib/parallel.rb#587
     def worker(job_factory, options, &block); end
   end
 end