--- a/test/unit/packwerk/parsers/erb_test.rb
+++ b/test/unit/packwerk/parsers/erb_test.rb
@@ -33,7 +33,7 @@ module Packwerk
         parser = stub
         parser.stubs(:ast).raises(err)
 
-        parser_class_stub = typed_mock(new: parser, "<=": true)
+        parser_class_stub = typed_mock(new: parser)
 
         parser = Erb.new(parser_class: parser_class_stub)
         file_path = fixture_path("invalid.erb")
@@ -54,7 +54,7 @@ module Packwerk
         parser = stub
         parser.stubs(:ast).raises(err)
 
-        parser_class_stub = typed_mock(new: parser, "<=": true)
+        parser_class_stub = typed_mock(new: parser)
 
         parser = Erb.new(parser_class: parser_class_stub)
         file_path = fixture_path("invalid.erb")