… Gemfile.lock
… sorbet/rbi/gems/actionpack@7.0.8.7.rbi
… sorbet/rbi/gems/actionview@7.0.8.7.rbi
… sorbet/rbi/gems/activesupport@7.0.8.7.rbi
… sorbet/rbi/gems/builder@3.3.0.rbi
… sorbet/rbi/gems/concurrent-ruby@1.3.5.rbi
… sorbet/rbi/gems/i18n@1.14.7.rbi
… sorbet/rbi/gems/json@2.7.2.rbi
… sorbet/rbi/gems/language_server-protocol@3.16.0.3.rbi
… sorbet/rbi/gems/language_server-protocol@3.17.0.3.rbi
… sorbet/rbi/gems/loofah@2.18.0.rbi
… sorbet/rbi/gems/loofah@2.24.0.rbi
… sorbet/rbi/gems/minitest@5.25.4.rbi
… sorbet/rbi/gems/mocha@2.5.0.rbi
… sorbet/rbi/gems/nokogiri@1.18.4.rbi
… sorbet/rbi/gems/parser@3.3.3.0.rbi
… sorbet/rbi/gems/prettier_print@0.1.0.rbi
… sorbet/rbi/gems/racc@1.8.1.rbi
… sorbet/rbi/gems/rack-test@2.2.0.rbi
… sorbet/rbi/gems/rack@2.2.13.rbi
… sorbet/rbi/gems/rails-dom-testing@2.0.3.rbi
… sorbet/rbi/gems/rails-dom-testing@2.2.0.rbi
… sorbet/rbi/gems/rails-html-sanitizer@1.4.3.rbi
… sorbet/rbi/gems/rails-html-sanitizer@1.6.2.rbi
… sorbet/rbi/gems/railties@7.0.8.7.rbi
… sorbet/rbi/gems/regexp_parser@2.9.2.rbi
… sorbet/rbi/gems/rexml@3.3.9.rbi
… sorbet/rbi/gems/rubocop-ast@1.31.3.rbi
… sorbet/rbi/gems/rubocop@1.64.1.rbi
… sorbet/rbi/gems/ruby-lsp@0.2.3.rbi