## What are you trying to accomplish?

Allow exclude patterns to filter load paths too. This is helpful when you have an embedded engine that is separate to your application, but still lives within the bounds of your root package.

## What approach did you choose and why?

Made a change in constant_resolver here: https://github.com/Shopify/constant_resolver/pull/44

I'll have to bump the minimum version when I release this one.

## What should reviewers focus on?

Before on a large app with several exclude patterns

```
bin/packwerk check  0.16s user 0.05s system 1% cpu 19.639 total
bin/packwerk check  0.16s user 0.06s system 1% cpu 19.285 total
bin/packwerk check  0.16s user 0.05s system 1% cpu 19.191 total
```

After on a large app with several exclude patterns

```
bin/packwerk check  0.16s user 0.05s system 1% cpu 19.770 total
bin/packwerk check  0.17s user 0.06s system 1% cpu 19.723 total
bin/packwerk check  0.16s user 0.05s system 1% cpu 19.447 total
```

Very slightly slower. Shouldn't impact performance much.

## Type of Change

- [x] Bugfix
- [ ] New feature
- [ ] Non-breaking change (a change that doesn't alter functionality - i.e., code refactor, configs, etc.)

### Additional Release Notes

- [ ] Breaking change (fix or feature that would cause existing functionality to change)

Include any notes here to include in the release description. For example, if you selected "breaking change" above, leave notes on how users can transition to this version.

If no additional notes are necessary, delete this section or leave it unchanged.

## Checklist

- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] It is safe to rollback this change.